+++ src/test/java/model/Ingredient.java
@@ -1,4 +1,4 @@
-package model.request;
+package model;
 
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
