+++ src/test/java/model/Ingredients.java
@@ -1,4 +1,4 @@
-package model.request;
+package model;
 
 
 import lombok.AllArgsConstructor;
