+++ src/test/java/model/response/IngredientResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import model.request.Ingredient;
+import model.Ingredient;
 
 import java.util.List;
