+++ src/test/java/model/response/OrdersListResponse.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import model.request.Order;
+import model.Order;
 
 import java.util.List;
