+++ src/test/java/test/order/OrderTestStep.java
@@ -4,8 +4,8 @@ import io.qameta.allure.Step;
 import io.restassured.http.Header;
 import io.restassured.http.Headers;
 import io.restassured.response.Response;
-import model.request.Ingredient;
-import model.request.Ingredients;
+import model.Ingredient;
+import model.Ingredients;
 import model.response.IngredientResponse;
 import org.apache.http.HttpStatus;
 import org.apache.http.protocol.HTTP;
