+++ src/test/java/test/user/CreateUserTest.java
@@ -3,7 +3,7 @@ package test.user;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
-import model.request.User;
+import model.User;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
