+++ src/test/java/test/user/LoginUserTest.java
@@ -3,7 +3,7 @@ package test.user;
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.response.Response;
-import model.request.User;
+import model.User;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
